requirements.txt
asgiref==3.5.0
asttokens==2.0.8
autopep8==1.6.0
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
entrypoints==0.4
executing==1.0.0
filelock==3.6.0
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
Markdown==3.3.7
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
tzdata==2022.1
virtualenv==20.14.1
wcwidth==0.2.5